Добавила штрафную функцию, числа с плавающей точкой

+++ hw1/memory.py
@@ -3,7 +3,7 @@ import numpy as np
 
 class MemoryVM:
     def __init__(self, size):
-        self.memory = np.zeros(size, dtype=np.int32)
+        self.memory = np.zeros(size, dtype=np.float32)
 
     def get(self, addr):
         return self.memory[addr]
